# apps/api-gateway/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/wines/{wineId}/similar": {
      "get": {
        "operationId": "WinesController_getSimilar",
        "parameters": [
          {
            "name": "wineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "wines"
        ]
      }
    },
    "/api/v1/wines/{wineId}": {
      "get": {
        "operationId": "WinesController_getById",
        "parameters": [
          {
            "name": "wineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "wines"
        ]
      }
    },
    "/api/v1/wines/submissions": {
      "post": {
        "operationId": "WinesController_submitWine",
        "summary": "Submit a wine to master library staging",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWineSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "wines"
        ]
      }
    },
...
    "/api/v1/wines/submissions/process": {
      "post": {
        "operationId": "WinesController_processSubmissions",
        "summary": "Process pending submissions (dedup worker)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessSubmissionsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "wines"
        ]
      }
    },
...
  },
  "info": {
    "title": "WineOps AI API",
    "description": "\n## Overview\n\nReal-time API for Restaurant Wine Inventory & Procurement Management.\n\n### Features\n\n- **Dashboard Aggregation**: Single endpoint for all dashboard data (API Bus pattern)\n- **One-Tap Actions**: Quick approval workflows for managers\n- **Toast POS Integration**: Proxy endpoints for Toast API\n- **Real-time Updates**: WebSocket support for live data sync\n- **Rate Limiting**: Per-user and per-restaurant quotas\n\n### Authentication\n\nAll endpoints (except public ones) require a Bearer token in the Authorization header.\n\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n### Rate Limits\n\n| Endpoint Type | Limit | Window |\n|--------------|-------|--------|\n| Default | 100 | 60s |\n| Auth | 10 | 60s |\n| Upload | 10 | 300s |\n| AI | 20 | 60s |\n\n### WebSocket Events\n\nConnect to `ws://localhost:4000` and join rooms:\n- `restaurant:{restaurantId}` - Restaurant-wide events\n- `user:{userId}` - User-specific events\n\nEvents:\n- `notification` - New notifications\n- `inventory_update` - Inventory changes\n- `order_update` - Order status changes\n- `one_tap_action` - One-tap action updates\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "WineOps Team",
      "url": "https://wineops.ai",
      "email": "support@wineops.ai"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://wineops.ai/license"
    }
  },
...
  "servers": [
    {
...
    },
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
...
      "CreateWineSubmissionDto": {
        "type": "object",
        "properties": {}
      },
      "ProcessSubmissionsDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}